Text description and synopsis
This project will encode or decode any text message using a key salted from a .PDF file.  The text message generates a string of digits which draw from a .PDF file by Title, Page and Pointer.  The title is shared before encoding.  The next pointer is a random seed generated using the title of the .PDF and an agreed upon modulo.  The first random number generated by the seed selects a page number in the .PDF.  The second number specifies the number of the character on the random page.  For example if the 23rd character on the page is equal to the first letter of the message then 23 could be used as the second number.   In cases where the second character does not occur in the page, then the second number will be a pointer to a space on the page and the third number will be a numeric counter signifying how many letters in alphabetic sequence beyond the character following the space that is the message letter.  In cases where the second character is on the page then the third number sequence is a random number generated from the system clock and the modulo.  In this manner each character of the message is represented by three numbers.